+++ docs/changelog.rst
@@ -13,6 +13,7 @@ Changelog
 - Internal error when an unresolvable pointer occurs during data generation.
 - Internal errors when references lead to non-objects.
 - Missing ``schema.override`` on schemas created via the ``from_pytest_fixture`` loader.
+- Not calling hooks for ``query`` / ``cookies`` / ``headers`` in GraphQL schemas. :issue:`1978`
 
 .. _v3.24.0: